requirements.txt
Django>=3.2.7,<3.3.0
djangorestframework>=3.12.4,<3.13.0
psycopg2>=2.7.5,<2.8.0
Pillow>=5.3.0,<5.4.0
python-dotenv>=0.19.0,<0.20.0
pytz==2021.1
asyncio>=3.4.3,<3.5.3
requests>=2.25.1,<2.26.0
aiohttp>=3.7.4.post0,<3.8.0.post0
futures3>=1.0.0,<1.1.0
PyJWT>=2.1.0,<2.2.0
Werkzeug>=2.0.1,<2.1.0
flake8>=3.9.2,<3.10.0
django-cors-headers==3.10.0

